Change in coredata requirements overlooked in other tests - now satisfied.
