+++ tests/__init__.py
@@ -7,7 +7,7 @@ options:
     nosetests --help
 
 """
-
+from datetime import datetime
 from flask_testing import TestCase as Base
 from flask_webtest import SessionScope
 
@@ -19,8 +19,10 @@ from portal.models.auth import Client
 from portal.models.fhir import CC
 from portal.models.fhir import add_static_concepts
 from portal.models.intervention import add_static_interventions, INTERVENTION
+from portal.models.organization import Organization
 from portal.models.relationship import add_static_relationships
 from portal.models.role import Role, add_static_roles, ROLE
+from portal.models.tou import ToU
 from portal.models.user import User, UserRoles
 
 TEST_USER_ID = 1
@@ -120,6 +122,21 @@ class TestCase(Base):
             self.test_user.save_constrained_observation(
                 codeable_concept=cc, value_quantity=CC.TRUE_VALUE,
                 audit=Audit(user_id=TEST_USER_ID))
+
+    def bless_with_basics(self):
+        """Bless test user with basic requirements for coredata"""
+        self.test_user.birthdate = datetime.utcnow()
+
+        # Register with a clinic
+        org = Organization(name='fake urology clinic')
+        self.test_user.organizations.append(org)
+
+        # Agree to Terms of Use
+        audit = Audit(user_id=TEST_USER_ID)
+        tou = ToU(audit=audit, text="filler text")
+        with SessionScope(db):
+            db.session.add(tou)
+            db.session.commit()
 
     def add_concepts(self):
         """Only tests needing concepts should load - VERY SLOW
